brake down into components
